Add files via upload
Files in which there are information about the events and more specifically we use them about the events, which are taking longer than a day. In these files we are saving an event, longer than a day, not as one event, but more. For example, if the event "Plan a Robbery" has a period 12/02/2023-14/02/2023, in these file we would have "Plan a Robbery" three times, so we can easy use them for choice 3 in the project.